# docs/swagger.yaml
definitions:
...
  util.CommonResult:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /flashSession/{id}:
    get:
      consumes:
      - application/json
      description: 获取场次详情
      parameters:
      - description: id
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 获取场次详情
      tags:
      - 限时购场次管理
  /flashSession/create:
    post:
      consumes:
      - application/json
      description: 添加场次
      parameters:
      - description: 创建时间
        in: query
        name: createTime
        type: string
      - description: 每日结束时间
        in: query
        name: endTime
        type: string
      - description: 编号
        in: query
        name: id
        type: integer
      - description: 场次名称
        in: query
        name: name
        type: string
      - description: 每日开始时间
        in: query
        name: startTime
        type: string
      - description: 启用状态：0->不启用；1->启用
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 添加场次
      tags:
      - 限时购场次管理
  /flashSession/delete/{id}:
    post:
      consumes:
      - application/json
      description: 删除场次
      parameters:
      - description: id
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 删除场次
      tags:
      - 限时购场次管理
  /flashSession/listAll:
    get:
      consumes:
      - application/json
      description: 获取全部场次
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 获取全部场次
      tags:
      - 限时购场次管理
  /flashSession/selectList:
    get:
      consumes:
      - application/json
      description: 获取全部可选场次及其数量
      parameters:
      - description: flashPromotionId
        in: query
        name: flashPromotionId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 获取全部可选场次及其数量
      tags:
      - 限时购场次管理
  /flashSession/update/{id}:
    post:
      consumes:
      - application/json
      description: 修改场次
      parameters:
      - description: id
        in: path
        name: id
        type: string
      - description: 创建时间
        in: query
        name: createTime
        type: string
      - description: 每日结束时间
        in: query
        name: endTime
        type: string
      - description: 编号
        in: query
        name: id
        type: integer
      - description: 场次名称
        in: query
        name: name
        type: string
      - description: 每日开始时间
        in: query
        name: startTime
        type: string
      - description: 启用状态：0->不启用；1->启用
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 修改场次
      tags:
      - 限时购场次管理
  /flashSession/update/status/{id}:
    post:
      consumes:
      - application/json
      description: 修改启用状态
      parameters:
      - description: id
        in: path
        name: id
        type: string
      - description: status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 修改启用状态
      tags:
      - 限时购场次管理
...
  /home/brand/create:
    post:
      consumes:
      - application/json
      description: 添加首页推荐品牌
      parameters:
      - in: query
        name: brandId
        type: integer
      - in: query
        name: brandName
        type: string
      - in: query
        name: id
        type: integer
      - in: query
        name: recommendStatus
        type: integer
      - in: query
        name: sort
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 添加首页推荐品牌
      tags:
      - 首页品牌管理
  /home/brand/delete:
    post:
      consumes:
      - application/json
      description: 批量删除推荐品牌
      parameters:
      - description: ids
        in: query
        items:
          type: string
        name: ids
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 批量删除推荐品牌
      tags:
      - 首页品牌管理
  /home/brand/list:
    get:
      consumes:
      - application/json
      description: 分页查询推荐品牌
      parameters:
      - description: brandName
        in: query
        name: brandName
        type: string
      - description: recommendStatus
        in: query
        name: recommendStatus
        type: string
      - description: 页码
        in: query
        name: pageNum
        type: string
      - description: 数量
        in: query
        name: pageSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 分页查询推荐品牌
      tags:
      - 首页品牌管理
  /home/brand/update/recommendStatus:
    post:
      consumes:
      - application/json
      description: 批量修改推荐品牌状态
      parameters:
      - description: ids
        in: query
        items:
          type: string
        name: ids
        type: array
      - description: recommendStatus
        in: query
        name: recommendStatus
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 批量修改推荐品牌状态
      tags:
      - 首页品牌管理
  /home/brand/update/sort/{id}:
    post:
      consumes:
      - application/json
      description: 修改推荐品牌排序
      parameters:
      - description: id
        in: path
        name: id
        type: string
      - description: sort
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.CommonResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.CommonResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.CommonResult'
      summary: 修改推荐品牌排序
      tags:
      - 首页品牌管理
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"